requirements.txt
streamlit
yfinance
pandas
numpy
plotly
redis
python-dotenv
bcrypt
ta-lib-binary
openai
python-dotenv